+++ x/distribution/types/expected_keepers.go
@@ -36,7 +36,7 @@ type BankKeeper interface {
 // PoolKeeper defines the expected interface needed to fund & distribute pool balances.
 type PoolKeeper interface {
 	FundCommunityPool(ctx context.Context, amount sdk.Coins, sender sdk.AccAddress) error
-	DistributeFromFeePool(ctx context.Context, amount sdk.Coins, receiveAddr sdk.AccAddress) error
+	DistributeFromCommunityPool(ctx context.Context, amount sdk.Coins, receiveAddr sdk.AccAddress) error
 	GetCommunityPool(ctx context.Context) (sdk.Coins, error)
 }